refactored authentication into feature folder
