+++ src/app/(authenticated)/editor/resume/[id]/page.tsx
@@ -1,5 +1,5 @@
 import styles from "./page.module.css";
-import { validateRequest } from "@/lib/auth";
+import { validateRequest } from "@/features/authentication/lib/auth";
 import { redirect } from "next/navigation";
 import { getResume } from "@/lib/resume";
 
@@ -24,7 +24,6 @@ const Page = async ({
 
     return (
         <main className={styles.main}>
-            <AddSectionModal />
             <PageUtilBar documentId={id} />
             <DocumentContainer document={resume} />
         </main>
