+++ src/app/(authentication)/email-verification/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 import styles from "./page.module.css";
-import Card from "@/components/authentication/email-verification-card/Card";
-import { validateRequest } from "@/lib/auth";
+import { EmailVerificationCard } from "@/features/authentication";
+import { validateRequest } from "@/features/authentication/lib/auth";
 
 const Page = async () => {
     const { user, session } = await validateRequest();
@@ -10,7 +10,7 @@ const Page = async () => {
 
     return (
         <main className={styles.pageContainer}>
-            <Card />
+            <EmailVerificationCard />
         </main>
     );
 };
