+++ src/app/(authentication)/password-reset/page.tsx
@@ -1,6 +1,6 @@
 import styles from "./page.module.css";
-import Card from "@/components/authentication/password-reset-email-card/Card";
-import { validateRequest } from "@/lib/auth";
+import { PasswordResetEmailCard } from "@/features/authentication";
+import { validateRequest } from "@/features/authentication/lib/auth";
 import { redirect } from "next/navigation";
 
 const PasswordResetPage = async () => {
@@ -10,7 +10,7 @@ const PasswordResetPage = async () => {
     }
     return (
         <main className={styles.pageContainer}>
-            <Card />
+            <PasswordResetEmailCard />
         </main>
     );
 };
