--- src/components/authentication/auth-nav/Nav.module.css
@@ -1,58 +1,0 @@
-.navContainer {
-    height: var(--nav-height);
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding-inline: var(--padding-inline-xlarge);
-    background-color: transparent;
-    z-index: 10;
-}
-
-.logoContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.textContainer {
-    height: var(--nav-height);
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding-inline: var(--padding-inline-large);
-    max-width: var(--max-width);
-}
-
-.companyName {
-    font-size: var(--font-size-medium);
-    font-weight: bold;
-}
-
-.navItem {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    gap: var(--gap-medium);
-}
-
-.navLink {
-    font-size: var(--font-size-medium);
-    color: black;
-}
-
-.reportButton {
-    background-color: white;
-    color: var(--accent-color);
-    font-weight: bold;
-    border-radius: var(--border-radius);
-    padding-inline: var(--padding-inline-small);
-    padding-block: var(--padding-block-small);
-    font-size: var(--font-size-small);
-    box-shadow: 0px 10px 10px 0px rgba(0, 0, 0, 0.1);
-    border: none;
-}
