+++ src/components/modals/payment/reinstate-modal/ReintstateModal.tsx
@@ -6,7 +6,7 @@ import { redirect } from "next/navigation";
 import Stripe from "stripe";
 import Link from "next/link";
 import { checkIcon, unlockIcon } from "@/components/icons/iconSVG";
-import { validateRequest } from "@/lib/auth";
+import { validateRequest } from "@/features/authentication/lib/auth";
 import { getUser } from "@/lib/user";
 import FormLoading from "@/components/form-loading/FormLoading";
