+++ src/features/authentication/types/auth.ts
@@ -1,0 +1,8 @@
+export type AuthEmailData = {
+    email: string;
+    firstName: string;
+    lastName: string;
+    subject: string;
+    url: string;
+    type: string;
+};
