+++ src/types.ts
@@ -1,12 +1,3 @@
-export type AuthEmailData = {
-    email: string;
-    firstName: string;
-    lastName: string;
-    subject: string;
-    url: string;
-    type: string;
-};
-
 export type FailedPaymentEmailData = {
     email: string;
     firstName: string;
